+++ src/Utility/ClassString/Exception/InvalidSubtypeClassNameException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Heptacom\HeptaConnect\Dataset\Base\Exception;
+namespace Heptacom\HeptaConnect\Utility\ClassString\Exception;
 
 class InvalidSubtypeClassNameException extends \InvalidArgumentException
 {